Update display methods to use BoxChars, leaving ASCII as a choice
I was successfully able to do this, but it took me a while. It's not that it was difficult, but rather, that there were a lot of little nuances, and also many ways to implement some of the edge case situations. I have it working exactly as I want it to now, though.

It handles even large grids fntastically, but once the tile count goes over 100, the format goes wonky because I have it set to two integer digits of display at all times. Though its unlikely anyone would want to play a game with over a 100 tiles, it would be fun to implement nonetheless. Its not a priority, but I see myself attempting it very soon anyway since its so goofy.

Next up is to implement the variable number of digits as mentioned just now, or implement a Range function/iterator, or implement tile swapping.