Update formatting and display methods to support arbitrary digit length
When I wrote this up until now, I had assumed the highest tile number would be 99 (well, 100, since the final number is not rendered, since it is a blank tile), so that the max number of rendered digits would be two. I thus created the formatting in a fixed manner to pad up to a max of two digits so that the grid would appear uniform. Well, it would be rare, but someone might actually want to play a game with over 100 tiles, and even if not, it would be fun and goofy to allow.

I thus implemented it in this commit. I have tested it up to values over 100k, and it works flawlessly. If I really want to get carried away with it, I could have it detect the dimensions of the terminal automatically and calculate max allowable grids, but that is premature optimization for something that might never get played by anyone other than myself.

It did occur to me during this commit that the tiles become quite rectangular as the digits of max tile numbers increase, and that it might look pleasant to write the formatting so that the tiles always remain 'approximately square'. But again, that would really be going overboard, for something that would better be implemented in a GUI of some sort (like SDK or Bitty Engine). I do have plans to do this (several plans actually - one for animated tiles and/or pictures on tiles, and another for machine learning/neural networks), so I feel it is unnecessary, though it would actually be interesting to try.

Next up is either implementing a functional Range function, or to get tile-swapping functioning. In the latter case, I should also implement optional toroidal geometry for the tile game.