Implement blank tile movement in a single direction
To get started with movement, I decided to implement it in a single direction first to see what the code would look like, before even attempting to generalize it so that a single method can perhaps use a map to do it in one place, for all types of movement.

I thus opted to start with moving 'up', which I have going. It stops when it hits the 'boundary', so toroidal geometry is not possible yet. It's also occurring to me that I really have the movement going in the direction opposite what would be expected, since when people play sliding tile puzzles, they don't see themselves as moving the empty tile, but rather as move the tiles adjacent to it into the blank space itself. Thus my current movement is unnatural.

This oversight by me is partly due to the fact that I'm still rendering the number of the 'blank tile'. I thus should implement the mechanism by which to blank it out.

Next up is to implement 'blanking out'/erasing the digits on the blank tile. After that, I should continue to implement movement in all directions, implement toroidal movement, and then generalize the movement methods into a single method (if such a thing is reasonable).