Actually 'blank-out' the blank tile
I updated the appropriate methods in order to make the highest-numbered tile blank, as is usually done in sliding tile games. The technique by which I did this is a bit sketchy, but I explain my logic in a comment right where it is implemented, detailing why it is a reasonable choice. Look there if interested (or in the git diff).

Next up is to continue implementing movement - finish the directions, correct it by inverting movement, implement toroidal movement, and consolidate it into a single method (but only if graceful and clean-looking).