Get toroidal movement working in one direction
In the existing moveUp method I have, I updated it so that toroidal movement is possible. REPL testing shows it works as expected. A simple commit, but I don't want to do too much in a single commit and would rather commit conservatively so I can charge ahead and jank things up if needed.

Next up is to implement movement in more directions, and see if it can be generalized nicely. I've decided to reverse the directions afterwards, since using the blank tile as the frame of reference is more intuitive for me while I design this.