Implement a string to function map for tile movement; implement movement in 2nd direction
Since I like to avoid ugly-looking if-chains and switch statements (which I know aren't avaiable in Lua, though it would be trivial to implement them via a generalized function), I've opted to implement a map that that maps expected player input strings to the corresponding methods for movement.

There was a bit of jankiness (which is explained in comments - it can be seen via the diff for this commit) which wrapping function placeholders for movement enums/constants, but there was a logic to it (I wanted to avoid forward declarations or forward implementations, and keep things in a logical order not mandated by the parser).

I also implemented the movement method for going downward. I've tested movement in both implemented directions - with and without toroidal geometry - as well as the general movement method that utilizes the map, via several different inputs at REPL - all worked gracefully. I can see a generalization pattern emerging, so I'll implement movement to the left and right, then once I see the totality of the movement functions, consolidate them into a single method, and then perhaps into the movement method currently using the mapper itself.

Next up is to finish the movement methods, then perhaps consolidate, and then invert the movements to match player expectations.