Implement remaining movement methods
I implemented movement methods for left and right, as well as corrected a bug whereby movement broke in toroidal geometry after 'wrapping-around' (I had forgotten to update the appropriate indices in such situations). Nothing fancy this commit. Committing because I might make some hefty changes if needed in the coming commits.

Next up is to look at the pattern amongst the four movement methods and consolidate them, if it makes sense. After that, it will be prudent to invert the movement methods to match player expectations.