Consolidate movement methods into a single method utilizing a lambda map; refactor swap method for readability
After looking at the totality of all the movement movements, I was able to generalize them into a single pattern (I've left the pseudocode in a comment, which I will remove next commit - I wanted to capture it in history, otherwise I would've already deleted it) using a second map which contained lambdas corresponding to the movement and needed behaviors.

I also refactored the swap method so that it didn't have to have four unsightly arguments passed in, and is now much nicer to look at.

After debugging two tiny bugs, everything is confirmed via REPL to be working wonderfully and exactly as desired. Both normal and toroidal movement are confirmed to be working correctly.

Next up should be to invert the movement to match traditional expectations regarding sliding tile puzzles, but I think I want to keep it uninverted now since that is more natural for me (maybe I could even make it a toggleable option, since its simply a matter of mapsthat its behavior is controlled - literally performing two swaps in the lambdaMap would invert the behavior as needed). Thus next up, I may want to write a method checking for a win condition, or a method that can scramble the tilegrid for the game (there is a specific way this needs to be done - impossible states are possible otherwise).