Start implementing a disordering function
I created a basic disordering function, and identified (but have not yet implemented) arguments that will be desirable to support. The disorder function has been tested at length via REPL and it works very well.

A bigger reason I want to commit right now is that I want to log some thoughts that occurred to me between now and the last commit, mostly regarding new ideas for other, new projects that are inspired by ideas I thought up here.

I believed I was going to have to implement something akin to an 'entropy' calculation here, to ascertain if puzzles had been 'shuffled' to a decent enough state to be nontrivial and play-worthy. This would involve doing something like calculating the Manhattan distance for each tile from where it needs to be, the dividing it by the highest possible Manhattan distance that it could possibly be in that puzzle (which it turns out, is not the same for each tile). This led me down a rabbit-hole of thoughts and ideas pertaining to combinatorics.

Things like 1) Can a maximum possible disordered-ness even be algorithmically calculated? 2) Can it be calculated at all, even non-algorithmically? 3) How does allowing toroidal movement affect such calculations and distances? 4) What about non-rectangular sliding puzzles, the can be convex or concave in nature? 5) How rapidly do different grids become disordered? 6) How does allowing any swaps, only adjacent swaps, and only the high-tile swap affect which disordered states can be reached, the maximum disordered-ness, and how rapidly a grid can become disordered? 7) How do different quantities of swaps, across the different categories of allowable swaps, affect disordered-ness? 7a) As one swaps into infinity, do cycles appear in the level of disordered-ness, and at what frequencies or intervals do they reoccur? 7b) Or does a disordered threshold get crossed, from which it is never gone under again? 8) How do the different types of swap restrictions and geometric movement affect the possibility / impossibility of the game, if the grid is shuffled under one but then to be solved in another (I know, for example, that arbitrary swaps can create impossible to solve puzzles if only the highest-numbered tile can be moved, but what about toroidal movement?)?

And so on, and so on... this has raised an almost endless number of questions for me, and I could keep on listing them for a while. It appears these questions have occurred to mathematicians before, especially in relation to linear sorting, but maybe less frequently in 2-d situations (and even more rarely in 3-d or higher-dimension situations), falling under graph theory and combinatorics. There are really so many cool and exploratory programs that could be made to examine these questions, even if adequate answers are never reached. The questions even blow up even further if you allow 'tiling' of different shapes, like hexagons.

Any, next up is to flesh out more the disorder function. Also, implementing move history in the TileGrid class would be nice, so that one can go backwards or see the solution at any time. It also is feeling more appropriate now to actually invert the movement as mentioned in previous commits. I could also just go ahead and implement the actual game loop now, since most of the core functionality is implemented. I'll do whatever I feel like in the moment when I get to it.