Implement move history tracking
I started implementing move history in TileGrid move functions. The tracking works, but a basic undo method I wrote needs to have an 'opposite map' implemented so that the reverse of the historical command can be sent to undo the previous movement. I also need to update Disorder so that disordering adds moves to the move history as well.

Next up is to make those few changes regarding move history made in the last few sentences.