Implement reverseMovementMap to enable proper move undo-ing
Very briefly implemented a reverseMovementMap, and then tested via REPL afterwards to ensure that undo'ing actually works as expected now. It does, in both normal and toroidal geometry.

Worth noting is that I had to introduce a parameter to the move function so that it can be indicated whether or not a move should be logged.

Unrelated to this but worth noting for later, is that I was mistaken earlier when I implemented the move constants as tables wrapping arbitrary values, in an attempt to stop duplication of data - it turns out, the Lua runtime is 'intelligent' enough to not duplicate strings when referenced multiples time, since they are immutable. Thus, if a string is written and it already exists somewhere, Lua just points a reference at that. It's not a big issue write now, especially since my code is working just fine, but it might be worth undoing that at some point and just letting the constants be strings instead of table-wrapped strings.

Next up, I should invert the movement, but I'll probably just start implementing the game interactive UI instead.