Port over inputGetter; start implementing game initializer
I spent most of this commit just fleshing out the general structure of the actual game, and then implementing the first decision that has to be made by the player (whether or not to play a game).

To do this, I ported over an input solicitation and validation function factory function I wrote in the past for a different Lua program, and adapted it for here by deleting unnecessary statements.

So far, everything I've tested works via REPL.

Next up is to actually implementing a round of game, and start implementing the core loop for it.