Implement another three inputGetters for game initializer
Beyond the inputGetter that determines whether a player wants to play a game or not from the previous commit, I implement three that solicit the height and width of the desired puzzle, as well as whether it is to be a toroidal puzzle or not. This involved updating the inputValidator logic to automatically convert the user input to a number if that is what is needed. Pretty trivial stuff.

It has all been tested and debugged via REPL until it worked exactly as desired, which it does not.

Next up is to either proceed with the core gameloops that solicits tile movements and displays them, or to implement a getter for the player's desired difficulty level, which can be crudely used to perform more swaps when shuffling the puzzle (which can be done intelligently or naively, depending on how complicated but accurate one wants it to be).