Get game playable - implement movement input getting and core game loop
The game is now functional in a minimal yet fully playable state. I implemented the inputGetter for tile movement inputs, and then a primary game loop that loops until the chosen puzzle is solved.

Testing via REPL shows it to be working as desired, and several games were played until they were solved.

Next up is to invert the slide movements as mentioned previously - it is quite disorienting right now, as they move opposite to what one would expect (this was intentional up until now). After that, 'undo' and 'quit' and 'solve' should be implemented as valid input options in the core game loop (among other unnamed input choices that might be desirable), as options available to the player other than movement. A 'help' input would be beneficial as well, to state in more detail how the game works and what the acceptable inputs are.