Invert movement to match player expectations; fix dimensions bug
This was a very trivial and brief commit - I simply used the reverse movement map to feed the opposite movement of the 'blank tile' into the move method for TileGame, effectively reversing the movement, so that it appears the player is controlling the numbered tiles adjacent to the blank tile, instead of the blank tile itself. It now moves as one would expect.

I also corrected a mistake where the dimensions solicited from the player were being fed mixed-up into the constructor.

Testing via REPL confirms this to be working as expected now, and it's actually a lot of fun playing it. I'm currently a third of the way through a 100-tile game.

Next up should be to implement other options besides movement during the core game loop, like undo, quit, solve, help, etc.