Implement branching needed for command input options; implement help option
I laid out and implemented the code needed to determine if a player-entered input is intended to be a tile movement or a command input, and then wrote out the skeletons for the related command functions those inputs would map to. I then implemented the help option.

After extensive testing and debugging to a careless oversight I made in my original implementation, I have it working exactly as desired, confirmed via REPL.

Next up is to implement the remaining input options - undo, solve, and quit. Quit would be a good starting point.