Implement quit puzzle input option; improve text spacing and formatting
I implemented a second input command option - 'quit' - which was quite trivial. I then went back across the various messages that are displayed and adjusted spacing and formatting to make it easier to read during play, and to make some of the messages more natural/correct.

Next up is to implement the solve or undo command inputs. I'm also seeing that it might be useful to implement functionality to track the move count for moves made by the player, as well an input command to display the current move count. This could also lead to a challenge mode (down the line - not now), where a puzzle can be generated and then the player must solve it in a finite amount of 'slides'.