Implement 'undo' command input during puzzle
Exactly as the title states. In order to stop the player from undoing the shuffling moves during the initialization of the puzzle, I had to implement the ability to track the number of moves made by the player (as opposed to simply retrieving such a count by obtaining the length of the move history). This is also useful, because displaying the number of moves via a command and at the end of the game would be informative.

The moves made by the disordering function should only be able to be undone during the game via the 'solve' input command.

Next up is to implement 'solve', but I am also seeing that 'reset' and 'movecount' input options would be beneficial. So at least one of those three should be implemented next.