Implemented option to reset puzzles to original state; implemented recursive table deep copier
I implemented the ability for players to reset the current puzzle to its original state (and this can be done as many times as they please), and the most direct and elegant method by which to do this occurred to me to be recursive deep table copying. It was straight-forward to implement, though I did have to do a little debugging before I got it going as I desired (careless typos).

I tested it pretty extensively via REPL, and it worked exactly as desired.

Next up is to implement the solve input option, and then this game is nearing completion (at least the terminal version of it).