Implement 'unintelligent' automatic puzzle solving
I implemented an input option that allows the player to have the program automatically solve the existing puzzle from wherever it currently is. It does this in an 'unintelligent' manner, by simply undoing the move history, including the moves made by the player and the moves made during the shuffling process.

I've tested it via REPL and it does indeed work and it works consistently. However, there are a few issues that could use addressing.

1) The shuffler currently does a flat 10k shuffles, which take ages to undo even at a fast rate. Large puzzles essentially never get solved in a timely fashion. This could be rememdied by having a disorder function that performs a number of swaps based on the size of the puzzle board. 2) Lots of moves/swaps are repeated back-to-back. This is because I never finished fully implementing the part of the disorder function which disallows moves that immediately undo a previous move. Going back and implementing that would make the solver much more interesting. 3) This disordering function would honestly benefit from some type of 'entropy'/measure of disorder calculation, after which a threshold is crossed, the disordering function stops and returns. This, however, is non-trivial, and I would need to study such measures via simulation for boards of various dimensions to see what kind of levels are reached on average, how many swaps it takes on average, how the levels oscillate over time, and so on (I detailed this in great length in a commit message many commits ago). That is probably beyond the scope of this particular program, but it is worth noting.

Next up is either implementing the restriction on repeated swaps, or some type of 'intelligent' means of determining a more reasonable number of swaps that are allowed during disordering. This program is actually in a releasable state right now.