Improve disordering function by preventing 'undoing' and 'illegal' moves; improve swap quantity calculation
As mentioned in my previous commits, I finally implemented the parameters needed in the disorder function to stop swaps that undo the immediately preceding swap, as well as stopped prevented moves from being counted as swaps and from incrementing the swap counter.

Additionally, I implemented a more 'intelligent' swap count determining calculation which is takens into account the dimensions of the grid as well as the theoretical maximum Manhattan distance.

The result is that running the 'solve' command during gameplay now looks much, MUCH better (as opposed to it just being the same swaps over and over again. All of the modified functionality was tested at length via REPL (had a lot of fun doing it too), and it all works exactly as desired.

Next up - well, this is effectively a complete project. I have accomplished all of the goals I set out to do for this. Documentation could be created, and a dockerfile could be created for those who want to try this without installing Lua. I'll get to all of that but it is not at all urgent or important to me. There are also a host of other features like multiplayer, co-op, multiplayer co-op and competition, AI versus, and online that would be fun to implement, but are better saved for GUI versions (which I fully intend to implement in another repo, perhaps even starting today). Also, I would love to study the entropy/measures of disorder that arise from the grid dimensions and swap behaviors over time, but that is an entirely separate, abstract project, that is of no pressing need, no matter how interesting it is to me (very). I'll do whatever comes to heart next, since there are so many paths forward.