Make move count displayable upon winning a puzzle; clean-up code and comments
As I was playing the game for fun, I was bothered by how I had no clue how many moves were made by me after I finished a game, which led to me run the 'moves' command on the second-to-last move each game so I could know. I thus updated the game to display the statistic whenever a game is won.

I then went back and cleaned up parts of the code that were bothering me - fixed typos, re-ordered some functions so that they were in a more natural order, and explained some parts of the code in greater detail.

Next up - this is effectively complete, as mentioned last commit. I plan on using it as the underlying logic for a visual game built on top of it, and may return here to update code only as it benefits the other projects I'm working on. May also return to create dockerfile and a release for the releases section, and to make official documentation.